Moved func to fix norm
